Rename Util to Resolver and change structure; Remove unused usings
